chore: move exporter related configurations into exporter mod.
